requirements.txt
fastapi
uvicorn
python-decouple
python-dotenv
pydantic[email]